feat: add login in auth service

+++ apps/api/src/modules/auth/auth.service.ts
@@ -1,17 +1,19 @@
-// src/auth/auth.service.ts
 import {
   ConflictException,
   Injectable,
   InternalServerErrorException,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { SupabaseService } from '../supabase/supabase.service';
 import { PrismaService } from '@packages/db';
-import { RegisterUserDto } from './dto/auth.dto';
+import { LoginUserDto, RegisterUserDto } from './dto/auth.dto';
+import { JwtService } from '@nestjs/jwt';
 @Injectable()
 export class AuthService {
   constructor(
     private readonly supabaseService: SupabaseService,
     private readonly prisma: PrismaService,
+    private readonly jwtService: JwtService,
   ) {}
 
   async registerUser(dto: RegisterUserDto) {
@@ -59,10 +61,50 @@ export class AuthService {
   async deleteUser(userId: string) {
     await this.supabaseService.deleteUser(userId);
 
-    await this.prisma.user.delete({
-      where: { id: userId },
-    });
+    if (!userId) {
+      throw new InternalServerErrorException('Supabase user deletion failed');
+    }
+
+    try {
+      await this.prisma.user.delete({
+        where: { id: userId },
+      });
+    } catch (error) {
+      throw new InternalServerErrorException(
+        `Failed to delete Supabase user: ${error.message}`,
+      );
+    }
 
     return { message: 'User deleted successfully' };
+  }
+
+  async login(dto: LoginUserDto) {
+    const { email, password } = dto;
+
+    const { data, error } =
+      await this.supabaseService.client.auth.signInWithPassword({
+        email,
+        password,
+      });
+
+    if (error || !data.user) {
+      throw new UnauthorizedException('Invalid credentials');
+    }
+
+    const user = await this.prisma.user.findUnique({
+      where: { id: data.user.id },
+    });
+
+    if (!user) {
+      throw new UnauthorizedException('User not found in DB');
+    }
+
+    const payload = { sub: user.id, email: user.email };
+    const token = this.jwtService.sign(payload);
+
+    return {
+      access_token: token,
+      user,
+    };
   }
 }
